+++ frontend/src/router/router.js
@@ -1,0 +1,13 @@
+import Vue from 'vue';
+import Router from 'vue-router';
+
+import routes from './routes';
+
+Vue.use(Router);
+
+const router = new Router({
+  mode: 'history',
+  routes,
+});
+
+export default router;
